### 2024-06-05 alchemistry/alchemlyb#376 @orbeckst
fix/update DOIs in JOSS paper draft
The jossbot found a number of missing DOIs https://github.com/openjournals/joss-reviews/issues/6845#issuecomment-2149351124 for the JOSS paper #328 .

We should start fixing our bibliography and adding _validated_ DOIs where possible.
