### 2024-06-05 alchemistry/alchemlyb#376 @orbeckst
From the Pre-REVIEW

MISSING DOIs

- 10.1016/0021-9991(76)90078-4 may be a valid DOI for title: Efficient estimation of free energy differences fr...
- No DOI given, and none found for title: Statistical mechanics of fluid mixtures
- No DOI given, and none found for title: Drug design: structure-and ligand-based approaches
- 10.1063/1.1740409 may be a valid DOI for title: High‐temperature equation of state by a perturbati...
- No DOI given, and none found for title: API design for machine learning software: experien...
- No DOI given, and none found for title: Simulation techniques for solvation-induced surfac...
- 10.1016/j.cpc.2021.108171 may be a valid DOI for title: LAMMPS - a flexible simulation tool for particle-b...
