+++ joss_paper/paper.md
@@ -216,7 +216,7 @@ Institutes of Health (Award No R01GM118772 to O.B., R35GM148236 to
 D.M., K08GM139031 to T.T.J.) and the National Science Foundation (award ACI-1443054 to O.B.). A.S. acknowledges funding by Deutsche Forschungsgemeinschaft (DFG, German Research Foundation) under Germany's Excellence Strategy - EXC 2075 – 390740016 and support by the Stuttgart Center for Simulation Science (SimTech).
 The sponsors were not involved in any aspects of the research or the writing of the manuscript.
 
-We thank Dominik Wille and Travis Jensen for substantial code contributions, Helmut Carter and We-Tse Hu for small fixes, Shujie Fan for initial code for fractional equilibration time calculation, and Jan Janssen for creating the initial conda-forge package.
+We thank Dominik Wille and Travis Jensen for substantial code contributions, Helmut Carter and Wei-Tse Hu for small fixes, Shujie Fan for initial code for fractional equilibration time calculation, and Jan Janssen for creating the initial conda-forge package.
 
 # Author contributions
 
