### 2024-06-05 e85a327
add missing DOIs
- fix #376
- added DOIs that were found by https://github.com/openjournals/joss-reviews/issues/6845#issuecomment-2149351124
