+++ joss_paper/paper.bib
@@ -6,6 +6,7 @@
   number={2},
   pages={245-268},
   year={1976},
+  doi={10.1016/0021-9991(76)90078-4},
 }
 
 
@@ -193,6 +194,7 @@
   number={8},
   pages={1420-1426},
   year={1954},
+  doi={10.1063/1.1740409},
 }
 
 @article{fan2020aa,
@@ -286,7 +288,8 @@
 	pages = 108171,
 	title = {{LAMMPS} - a flexible simulation tool for particle-based materials modeling at the atomic, meso, and continuum scales},
 	volume = 271,
-	year = 2022
+	year = 2022,
+	doi = {10.1016/j.cpc.2021.108171},
 }
 
 
