+++ joss_paper/paper.bib
@@ -89,6 +89,7 @@
   number={5},
   pages={300-313},
   year={1935},
+  doi={10.1063/1.1749657},
 }
 
 @article{klimovich2015guidelines,
