### 2024-06-05 alchemistry/alchemlyb#376 @orbeckst
fixed with e85a327eb28342a2b95e0d118981770083372e95
