### 2024-06-10 22d72ac
Sort lambda values for consistent output
