### 2024-06-10 b2c22e4
Sort lambda values for consistent output
