### 2024-06-17 #6845 @srmnitc
👋🏽 @oostenbrink @philbiggin, would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Thanks!
