### 2024-06-27 #6845 @philbiggin
I can do this, but do I have to operate through GitHub?

If you send a standard email I am more likely to pick things up ;-)

Best wishes,
Phil

--------------------------------------------
Prof Phil Biggin
T:  01865 613305
W: sbcb.bioch.ox.ac.uk/biggin.php
E: ***@***.***
