### 2024-06-27 #6934 @editorialbot
[REVIEW]: alchemlyb: the simple alchemistry library
**Submitting author:** @orbeckst (<a href="http://orcid.org/0000-0003-1340-0831">Oliver Beckstein</a>)
**Repository:** https://github.com/alchemistry/alchemlyb
**Branch with paper.md** (empty if default branch): 71-joss-paper
**Version:** 2.4.1
