### 2024-06-27 #6934 @philbiggin
If this is an issue (ie publishing with one of the authors in the last four years) then I would also have to withdrawn
Best wishes,
Phil

--------------------------------------------
Prof Phil Biggin
T:  01865 613305
W: sbcb.bioch.ox.ac.uk/biggin.php
E: ***@***.***
