### 2024-06-28 #6934 @mikemhenry
@philbiggin This is the policy https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy fyi
