### 2024-06-28 #6934 @philbiggin
In that case I am definitely conflicted as having published with several of the authors.

Phil

-----------------------------------------------------------
Prof Phil Biggin
Professor of Computational Biochemistry
Chair of the Molecular Graphics and Modelling Society
Fellow and Tutor at Lady Margaret Hall, Oxford
Department of Biochemistry
South Parks Road
Oxford
OX1 3QU
E: ***@***.***
W: sbcb.bioch.ox.ac.uk/biggin.php
T:  +44 1865 613305
