### 2024-07-03 f1d33dd
Update matplotlib depdendency
