+++ environment.yml
@@ -9,5 +9,5 @@ dependencies:
 - scipy
 - scikit-learn
 - pyarrow
-- matplotlib
+- matplotlib>=3.7
 - loguru
