### 2024-07-04 5a73cc4
Support matplotlib 3.9.0 (#377)
* fix failure related to outdated matplotlib usage (legend_handles)
* fix alchemistry/flamel#28
* update CHANGES
