+++ devtools/conda-envs/test_env.yaml
@@ -8,7 +8,7 @@ dependencies:
 - pymbar>=4
 - scipy
 - scikit-learn
-- matplotlib
+- matplotlib>=3.7
 - loguru
 - pyarrow
 
