### 2024-07-06 f3c15b7
reformat the rst readme as md
