+++ setup.py
@@ -41,7 +41,7 @@ setup(
     packages=find_packages("src"),
     package_dir={"": "src"},
     license="BSD",
-    long_description=open("README.rst").read(),
+    long_description=open("README.md").read(),
     long_description_content_type="text/x-rst",
     python_requires=">=3.9",
     tests_require=["pytest", "alchemtest"],
