+++ setup.py
@@ -42,7 +42,7 @@ setup(
     package_dir={"": "src"},
     license="BSD",
     long_description=open("README.md").read(),
-    long_description_content_type="text/x-rst",
+    long_description_content_type="text/markdown",
     python_requires=">=3.9",
     tests_require=["pytest", "alchemtest"],
     install_requires=[
