+++ MANIFEST.in
@@ -1,6 +1,6 @@
 include *.rst
 recursive-include src/alchemlyb *.gz *.bz2 *.zip *.rst *.txt
-include README.rst
+include README.md
 include LICENSE
 include versioneer.py
 include src/alchemlyb/_version.py
