### 2024-07-08 f40b89b
Prepare 2.3.1 (#378)
* reformat the rst readme as md

* update setup.py

* address comments
