### 2024-07-15 23bc5c5
Update error for lambda values
