### 2024-07-15 d8e7092
Update extract_dHdl* to use prec
