### 2024-07-24 #6934 @glycodynamics
@srmnitc It is not related to code. The code worked perfectly fine for the test case I tried. My comments are related to the "State of the field" and "Example usage" sections, which I did not check off in my checklist.

- Discussion of similar tools would be helpful for context and it's missing in the manuscript.
- Although documentation is good, providing a single example case with input files would greatly benefit novice users in understanding the process.
- I noticed that the description of errors is missing from the paper, which could be a valuable addition.

For editor only: I noticed some co-authors contributed less code and commits than one who has been just acknowledged. Also, at least one co-author is not listed in the copyright section of the code but is listed as a co-author in the paper. I understand that the contributions of each author are best known to the corresponding author and not raising a flag but pointing it out as there was a "Contribution and authorship" section in the checklist.
