### 2024-07-24 #6934 @ryankzhu
@srmnitc I am sorry for the delay. The codes have worked fine in my test case. The documentation is quite detailed and the manuscript reads nicely and concisely to me. I only have two small comments:
* It might be good to have a line (in addition to the badge) in the README to link to the documentation to make it more noticeable.
* I totally agree with @glycodynamics that it would be useful to have a simple example, e.g., a self-contained jupyter notebook, to walk through the main functionalities.
