### 2024-07-24 #6934 @orbeckst
Thank you for your reviews @glycodynamics and @ryankzhu .

Would you be able to open issues in https://github.com/alchemistry/alchemlyb/issues for
* providing a single example case with input file, e.g. simple example, e.g., a self-contained jupyter notebook, to walk through the main functionalities.
* add link to docs inside README

We will address @glycodynamics 's comments in the paper
* Discussion of similar tools would be helpful for context and it's missing in the manuscript.
  * _We do mention alchemical-analysis.py ; I am not aware of other tools that do the same as alchemlyb in terms of being applicable to many different file format, but we'll look carefully. If you have any specific software in mind please let us know!_
* I noticed that the description of errors is missing from the paper, which could be a valuable addition.
  * _Do you mean estimation of errors?_

cc first author @xiki-tempula
