### 2024-07-24 #6934 @orbeckst
> I noticed some co-authors contributed less code and commits than one who has been just acknowledged. Also, at least one co-author is not listed in the copyright section of the code but is listed as a co-author in the paper. I understand that the contributions of each author are best known to the corresponding author and not raising a flag but pointing it out as there was a "Contribution and authorship" section in the checklist.

The PR https://github.com/alchemistry/alchemlyb/pull/328 contains the full history of who decided to be an author and full rationale for all decisions regarding authorship. I am happy to explain in more detail.
