### 2024-07-24 #6934 @glycodynamics
> We will address @glycodynamics 's comments in the paper
> * Discussion of similar tools would be helpful for context and it's missing in the manuscript.
>
>   * _We do mention alchemical-analysis.py ; I am not aware of other tools that do the same as alchemlyb in terms of being applicable to many different file format, but we'll look carefully. If you have any specific software in mind please let us know!_
> * I noticed that the description of errors is missing from the paper, which could be a valuable addition.
>
>   * _Do you mean estimation of errors?_
>
> cc first author @xiki-tempula

@orbeckst
- I agree that no other tools can handle output from multiple programs. It may be worth mentioning analysis tools of MD programs if they offer the same estimators and provide functionalities like checking convergence, error estimates, and overlap matrix. For example, analyse_freenrg from siremol is the one that came to my mind when I was testing and saw the plot of the overlap matrix.

- I meant to say estimation of errors, how are they obtained, internally in alchemylib or coming from the external program, such as bootstrapping in MBAR? I may have missed, but I didn't find good documentation of this.

Thanks and best wishes.
