### 2024-07-24 809e0c6
Add moving_average function for visualization and convergence testing
