+++ src/alchemlyb/visualisation/convergence.py
@@ -136,11 +136,133 @@ def plot_convergence(dataframe, units=None, final_error=None, ax=None):
     ax.legend(
         (line1[0], line2[0]),
         ("Forward", "Reverse"),
-        loc=9,
+        loc="best",
         prop=FP(size=18),
         frameon=False,
     )
     ax.set_xlabel(r"Fraction of the simulation time", fontsize=16, color="#151B54")
+    ax.set_ylabel(r"$\Delta G$ ({})".format(units), fontsize=16, color="#151B54")
+    plt.tick_params(axis="x", color="#D2B9D3")
+    plt.tick_params(axis="y", color="#D2B9D3")
+    return ax
+
+def plot_moving_average(dataframe, units=None, final_error=None, ax=None):
+    """Plot the forward and backward convergence.
+
+     The input could be the result from
+     :func:`~alchemlyb.convergence.forward_backward_convergence` or
+     :func:`~alchemlyb.convergence.fwdrev_cumavg_Rc`. The input should be a
+     :class:`pandas.DataFrame` which has column `FE` and
+     :attr:`pandas.DataFrame.attrs` should compile with :ref:`note-on-units`.
+     The errorbar will be plotted if column `FE_Error` and `Backward_Error`
+     is present.
+
+     `FE`: A column of free energy estimate from some X% block of the data,
+     where optional `FE_Error` column is the corresponding error.
+
+    `final_error` is the error of the final value and is shown as the error band around the
+    final value. It can be provided in case an estimate is available that is more appropriate
+    than the default, which is the error of the last value in `Backward`.
+
+     Parameters
+     ----------
+     dataframe : Dataframe
+         Output Dataframe has column `Forward`, `Backward` or optionally
+         `Forward_Error`, `Backward_Error` see :ref:`plot_convergence <plot_convergence>`.
+     units : str
+         The unit of the estimate. The default is `None`, which is to use the
+         unit in the input. Setting this will change the output unit.
+     final_error : float
+         The error of the final value in ``units``. If not given, takes the overall
+         error of the time blocks, unless these were not provided, it which case it
+         equals 1 kT.
+     ax : matplotlib.axes.Axes
+         Matplotlib axes object where the plot will be drawn on. If ``ax=None``,
+         a new axes will be generated.
+
+     Returns
+     -------
+     matplotlib.axes.Axes
+         An axes with the forward and backward convergence drawn.
+
+     Note
+     ----
+     The code is taken and modified from
+     `Alchemical Analysis <https://github.com/MobleyLab/alchemical-analysis>`_.
+
+
+     .. versionchanged:: 1.0.0
+         Keyword arg final_error for plotting a horizontal error bar.
+         The array input has been deprecated.
+         The units default to `None` which uses the units in the input.
+
+     .. versionchanged:: 0.6.0
+         data now takes in dataframe
+
+     .. versionadded:: 0.4.0
+    """
+    if units is not None:
+        dataframe = get_unit_converter(units)(dataframe)
+    df_avg = dataframe["FE"].to_numpy()
+    if "FE_Error" in dataframe:
+        df_avg_error = dataframe["FE_Error"].to_numpy()
+    else:
+        df_avg_error = np.zeros(len(df_avg))
+
+    if ax is None:  # pragma: no cover
+        fig, ax = plt.subplots(figsize=(8, 6))
+
+    plt.setp(ax.spines["bottom"], color="#D2B9D3", lw=3, zorder=-2)
[…]
