### 2024-07-24 #6934 @xiki-tempula
@glycodynamics Thanks for the review. I'm glad that you mentioned the free energy analysis tool in SOMD. So analyse_freenrg from siremol is deprecated and their more recent tool for doing free energy analysis is based on alchemlyb https://github.com/OpenBioSim/biosimspace/blob/devel/python/BioSimSpace/FreeEnergy/_relative.py#L1241.

> I meant to say estimation of errors, how are they obtained, internally in alchemylib or coming from the external program, such as bootstrapping in MBAR? I may have missed, but I didn't find good documentation of this.

We just use the error estimation function from pymbar.
