+++ src/alchemlyb/visualisation/convergence.py
@@ -190,16 +190,7 @@ def plot_moving_average(dataframe, units=None, final_error=None, ax=None):
      The code is taken and modified from
      `Alchemical Analysis <https://github.com/MobleyLab/alchemical-analysis>`_.
 
-
-     .. versionchanged:: 1.0.0
-         Keyword arg final_error for plotting a horizontal error bar.
-         The array input has been deprecated.
-         The units default to `None` which uses the units in the input.
-
-     .. versionchanged:: 0.6.0
-         data now takes in dataframe
-
-     .. versionadded:: 0.4.0
+     .. versionadded:: 2.4.0
     """
     if units is not None:
         dataframe = get_unit_converter(units)(dataframe)
