### 2024-07-24 809c6f0
update parsing.lammps.py docstrings
