+++ src/alchemlyb/parsing/lammps.py
@@ -14,7 +14,7 @@ approximation of the Hamiltonian) at specified values of :math:`\lambda` and :ma
 LAMMPS, `fix adapt/fep <https://docs.lammps.org/fix_adapt_fep.html>`_ changes :math:`\lambda` and 
 `compute fep <https://docs.lammps.org/compute_fep.html>`_ changes :math:`\lambda'`.
 
-.. versionadded:: 1.0.0
+.. versionadded:: 2.4.0
 
 """
 
@@ -40,6 +40,8 @@ def _isfloat(x):
 def beta_from_units(T, units):
     """Output value of beta from temperature and units.
 
+    Supported types are: cgs, electron, lj. metal, micro, nano, real, si
+
     Parameters
     ----------
     T : float
@@ -57,14 +59,14 @@ def beta_from_units(T, units):
     ValueError
         If unit string is not recognized.
 
-    .. versionadded:: 1.??
+    .. versionadded:: 2.4.0
     """
     if units == "real":  # E in kcal/mol, T in K
         beta = 1 / (R_kJmol * kJ2kcal * T)
     elif units == "lj":  # Nondimensional E and T scaled by epsilon
         beta = 1 / T
     elif units == "metal":  # E in eV, T in K
-        beta = 1 / (R_kJmol * kJ2kcal * T)  # NoteHere!!!!
+        beta = 1 / (R_kJmol * kJ2kcal * T)
     elif units == "si":  # E in J, T in K
         beta = 1 / (constants.R * T * constants.physical_constants["electron volt"][0])
     elif units == "cgs":  # E in ergs, T in K
@@ -77,7 +79,8 @@ def beta_from_units(T, units):
         beta = 1 / (constants.R * T * 1e-21)
     else:
         raise ValueError(
-            "LAMMPS unit type, {}, is not supported. Supported types are: real and lj".format(
+            "LAMMPS unit type, {}, is not supported. Supported types are: cgs, electron,"
+            " lj. metal, micro, nano, real, si".format(
                 units
             )
         )
@@ -104,18 +107,23 @@ def _tuple_from_filename(filename, separator="_", indices=[2, 3], prec=4):
     tuple[float]
         Tuple of lambda values
 
-    .. versionadded:: 1.??
+    .. versionadded:: 2.4.0
     """
 
     name_array = ".".join(os.path.split(filename)[-1].split(".")[:-1]).split(separator)
-    if not _isfloat(name_array[indices[0]]):
+    try:
+        float(name_array[indices[0]])
+    except ValueError:
         raise ValueError(
-            f"Entry, {indices[0]} in filename cannot be converted to float: {name_array[indices[0]]}"
-        )
-    if not _isfloat(name_array[indices[1]]):
+                f"Entry, {indices[0]} in filename cannot be converted to float: {name_array[indices[0]]}"
+        (
+    try:
+        float(name_array[indices[1]])
+    except ValueError:
         raise ValueError(
-            f"Entry, {indices[1]} in filename cannot be converted to float: {name_array[indices[1]]}"
+                f"Entry, {indices[1]} in filename cannot be converted to float: {name_array[indices[1]]}"
         )
+
     return (
         round(float(name_array[indices[0]]), prec),
         round(float(name_array[indices[1]]), prec),
@@ -144,10 +152,12 @@ def _lambda_from_filename(filename, separator="_", index=-1, prec=4):
     float
         Lambda prime value
 
-    .. versionadded:: 1.??
+    .. versionadded:: 2.4.0
     """
     name_array = ".".join(os.path.split(filename)[-1].split(".")[:-1]).split(separator)
-    if not _isfloat(name_array[index]):
+    try:
+        float(name_array[index])
+    except:
[…]
