+++ docs/visualisation/alchemlyb.visualisation.plot_moving_average.rst
@@ -1,0 +1,6 @@
+﻿alchemlyb.visualisation.plot\_moving\_average
+=============================================
+
+.. currentmodule:: alchemlyb.visualisation
+
+.. autofunction:: plot_moving_average
