+++ docs/convergence.rst
@@ -75,6 +75,31 @@ is, where 0 fully-unequilibrated and 1.0 is fully-equilibrated. ::
     >>> value = A_c(dhdl_list, tol=2)
     0.7085
 
+Moving Average
+--------------
+If one obtains suspicious results from the forward / backward convergence plot,
+it may be useful to view the moving average of the change in free energy using
+:func:`~alchemlyb.convergence.moving_average` and
+:func:`~alchemlyb.visualisation.plot_moving_average` over the course of each 
+step in lambda individually, the following example is for :math:`\lambda` = 0
+
+    >>> from alchemtest.gmx import load_benzene
+    >>> from alchemlyb.parsing.gmx import extract_u_nk
+    >>> from alchemlyb.visualisation import plot_moving_average
+    >>> from alchemlyb.convergence import moving_average
+
+    >>> bz = load_benzene().data
+    >>> data_list = [extract_u_nk(xvg, T=300) for xvg in bz['Coulomb']]
+    >>> df = moving_average(data_list, 'mbar')
+    >>> ax = plot_moving_average(df)
+    >>> ax.figure.savefig('dF_t_moving_average.png')
+
+Will give a plot looks like this
+
+.. figure:: images/dF_t_moving_average.png
+
+   A convergence plot of showing that the forward and backward has converged
+   fully.
 
 Convergence functions
 ---------------------
