### 2024-07-24 alchemistry/alchemlyb#382 @ryankzhu
[JOSS] Adding tutorials and link to documentation
Hi alchemlyb team,

This is just to organize the comments from https://github.com/openjournals/joss-reviews/issues/6934 related to the repo:

* Add an example case with input files to walk through the main functionalities
* Add a link to docs inside README

Thanks and best wishes.
