### 2024-07-25 #6934 @srmnitc
@glycodynamics and @ryankzhu  thanks for the review, and thanks @orbeckst and @xiki-tempula for agreeing to address the issues.

@glycodynamics Thanks again for looking closely at the author list. @orbeckst going through https://github.com/alchemistry/alchemlyb/pull/328 it looks like the author list is justified. However, as @glycodynamics pointed out there are differences between copyright statement on the webpage and the author list. Is there a reason for this difference? Would it make more sense to be consistent?
