### 2024-07-25 b55552f
Formatting to align with Black
