+++ src/alchemlyb/estimators/bar_.py
@@ -88,21 +88,22 @@ class BAR(BaseEstimator, _EstimatorMixOut):
         # sort by state so that rows from same state are in contiguous blocks
         u_nk = u_nk.sort_index(level=u_nk.index.names[1:])
 
+        # get a list of the lambda states that are sampled
+        self._states_ = u_nk.columns.values.tolist()
+
         # group u_nk by lambda states
         groups = u_nk.groupby(level=u_nk.index.names[1:])
         N_k = [
             (len(groups.get_group(i)) if i in groups.groups else 0)
             for i in u_nk.columns
         ]
-        
-        # get a list of the lambda states that are sampled
-        self._states_ = [x for i, x in enumerate(u_nk.columns.values.tolist()) if N_k[i] > 0]
-        N_k = [x for x in N_k if x > 0]
-
+        states = [x for i, x in enumerate(self._states_) if N_k[i] > 0]
         # Now get free energy differences and their uncertainties for each step
         deltas = np.array([])
         d_deltas = np.array([])
         for k in range(len(N_k) - 1):
+            if N_k[k] == 0 or N_k[k + 1] == 0:
+                continue
             # get us from lambda step k
             uk = groups.get_group(self._states_[k])
             # get w_F
@@ -110,7 +111,7 @@ class BAR(BaseEstimator, _EstimatorMixOut):
 
             # get us from lambda step k+1
             uk1 = groups.get_group(self._states_[k + 1])
-            
+
             # get w_R
             w_r = uk1.iloc[:, k] - uk1.iloc[:, k + 1]
 
@@ -152,13 +153,11 @@ class BAR(BaseEstimator, _EstimatorMixOut):
             ad_delta += np.diagflat(np.array(dout), k=j + 1)
 
         # yield standard delta_f_ free energies between each state
-        self._delta_f_ = pd.DataFrame(
-            adelta - adelta.T, columns=self._states_, index=self._states_
-        )
+        self._delta_f_ = pd.DataFrame(adelta - adelta.T, columns=states, index=states)
 
         # yield standard deviation d_delta_f_ between each state
         self._d_delta_f_ = pd.DataFrame(
-            np.sqrt(ad_delta + ad_delta.T), columns=self._states_, index=self._states_
+            np.sqrt(ad_delta + ad_delta.T), columns=states, index=states
         )
         self._delta_f_.attrs = u_nk.attrs
         self._d_delta_f_.attrs = u_nk.attrs
