### 2024-07-26 124f148
Add a tutorial section (#379)
* fix #382
* address some of the reviewer comments inSee openjournals/joss-reviews#6934
  -  add link to docs in README
   - add tutorial section to docs
* update CHANGES

---------

Co-authored-by: Oliver Beckstein <orbeckst@gmail.com>
