### 2024-07-27 alchemistry/alchemlyb#383 @orbeckst
[JOSS] inconsistent author list between docs and AUTHORS
As pointed out in https://github.com/openjournals/joss-reviews/issues/6934#issuecomment-2249664394 , the authors in the doc footer

![image](https://github.com/user-attachments/assets/561f6648-7705-4973-81d1-7ecc18abe208)

do not include all the authors in [AUTHORS](https://github.com/alchemistry/alchemlyb/blob/master/AUTHORS), which is the authoritative reference for all code contributors.

We should either include everyone in the doc footer (generate the list for the docs programmatically) or simply link to the AUTHORS file.
