### 2024-07-27 #6934 @orbeckst
We should be consistent between AUTHORS and doc footer and I raised an issue to address this discrepancy.
