### 2024-07-27 alchemistry/alchemlyb#383 @orbeckst
I am in favor of just stating as authors "The [alchemlyb development team](https://github.com/alchemistry/alchemlyb/blob/master/AUTHORS)." with a link to AUTHORS.
