### 2024-07-27 alchemistry/alchemlyb#383 @xiki-tempula
@orbeckst I do see the point as it will be painful to maintain a up-to-date AUTHORS. But how shall we fix this? Sorry the suggestion is not too clear to me.
