### 2024-07-27 alchemistry/alchemlyb#383 @orbeckst
I'll have a look, I've done something similar for other projects.--Oliver Beckstein * ***@***.***://becksteinlab.physics.asu.eduOn Jul 27, 2024, at 12:01 PM, Zhiyi Wu ***@***.***> wrote:﻿
@orbeckst I do see the point as it will be painful to maintain a up-to-date AUTHORS. But how shall we fix this? Sorry the suggestion is not too clear to me.
