### 2024-07-28 da4e8ac
replace doc footer
- fix #383
- replaced explicit (incomplete) author list in doc footer with
  The alchemlyb development team and a link to AUTHORS
- automatically insert the current year into copyright
- Needed to hack the RTD theme to avoid the URL escape on the copyright
  variable by including a _templates/footer.html template.
