### 2024-07-28 #6934 @orbeckst
We fixed the inconsistent author display in the docs by linking to the authoritative AUTHORS file. Please see cross-linked issue alchemistry/alchemlyb#383 .
