### 2024-07-29 #6934 @glycodynamics
@orbeckst Thanks so much for the revision.

@srmnitc I read both the tutorial and how authors are now handling the list of developers. I agree that mentioning about the error estimation function is not required unless it was handled somehow differently within the alchemlyb. I am happy with the revision and I do not have any further comments.
