### 2024-08-15 alchemistry/alchemlyb#386 @srmnitc
Changes to paper
This is part of review over at https://github.com/openjournals/joss-reviews/issues/6934

Small small items from my side. Feel free to choose as needed.

- Shuai Liu is missing ORCid
- Affiliations: some of the have zip codes, some of them dont
- line 23-25 could be better with a general citation
- line 81-84 could be better with a general citation
- line 111 - you could remove the word 'doing'
- line 182 - is it un-equilibrated or non-equilibrated?
- line 198 - formatting for 'alchemlyb' is not there
- line 221 - citation, please check DOI
- line 257 - citation, please recheck DOI
- line 271 - citation, please recheck DOI
- line 293 - citation, please recheck DOI
- some journal names are abbreviations, some are full, could you please make them consistent?

Thanks a lot!
