### 2024-08-15 alchemistry/alchemlyb#386 @orbeckst
@xiki-tempula do you have time to address the requested changes? I am at workshops/conferences/travelling until end of August and only have _very_ limited bandwidth.
