### 2024-08-15 alchemistry/alchemlyb#386 @xiki-tempula
I think I might have some time at the weekends.
