+++ joss_paper/paper.md
@@ -69,7 +69,7 @@ affiliations:
    index: 2
  - name: Datryllic LLC, Phoenix, Arizona, USA (present affiliation)
    index: 3
- - name: Open Free Energy, Open Molecular Software Foundation, Davis, 95616 California, United States
+ - name: Open Free Energy, Open Molecular Software Foundation, Davis, California, United States
    index: 4
  - name: Differentiated Therapeutics, San Diego, CA
    index: 5
@@ -91,7 +91,7 @@ affiliations:
    index: 13
  - name: PM Scientific Consulting, Basel, Switzerland
    index: 14
- - name: Stuttgart Center for Simulation Science (SC SimTech) & Institute for Computational Physics, University of Stuttgart, 70569 Stuttgart, Germany 
+ - name: Stuttgart Center for Simulation Science (SC SimTech) & Institute for Computational Physics, University of Stuttgart, Stuttgart, Germany 
    index: 15
  - name: University of Colorado Boulder, Boulder, Colorado, USA
    index: 16
@@ -105,7 +105,7 @@ bibliography: paper.bib
 
 # Summary
 
-*alchemlyb* is an open-source Python software package for the analysis of alchemical free energy calculations, an important method in computational chemistry and biology, most notably in the field of drug discovery.
+*alchemlyb* is an open-source Python software package for the analysis of alchemical free energy calculations, an important method in computational chemistry and biology, most notably in the field of drug discovery [@merz2010drug].
 Its functionality contains individual composable building blocks for all aspects of a full typical free energy analysis workflow, starting with the extraction of raw data from the output of diverse molecular simulation packages, moving on to data preprocessing tasks such as decorrelation of time series, using various estimators to derive free energy estimates from simulation samples, and finally providing quality analysis tools for data convergence checking and visualization.
 *alchemlyb* also contains high-level end-to-end workflows that combine multiple building blocks into a user-friendly analysis pipeline from the initial data input stage to the final results. This workflow functionality enhances accessibility by enabling researchers from diverse scientific backgrounds, and not solely computational chemistry specialists, to use *alchemlyb* effectively.
 
@@ -140,7 +140,7 @@ Notably, *alchemlyb*'s robust and user-friendly nature has led to its integratio
 
 # Background: Alchemical free energy calculations
 
-Free energy differences are fundamental to understand many different processes at the molecular scale, ranging from the binding of drug molecules to their receptor proteins or nucleic acids through the partitioning of molecules into different solvents or phases to the stability of crystals and biomolecules.
+Free energy differences are fundamental to understand many different processes at the molecular scale, ranging from the binding of drug molecules to their receptor proteins or nucleic acids through the partitioning of molecules into different solvents or phases to the stability of crystals and biomolecules [@deng2009computations].
 The calculation of such transfer free energies involves constructing two end states where a target molecule interacts with different environments.
 For example, in a solvation free energy calculation, in one state (the coupled state) the target molecule interacts with a solvent (in the case of hydration free energies, water), while in the other state (the decoupled state) the ligand has no intermolecular interactions, which mimics the transfer of a ligand from infinite dilution in the solvent to the gas phase.
 The solvation free energy is then obtained by calculating the free energy difference between these two end states, but it is crucial to ensure sufficient overlap in phase space between the coupled and decoupled states, a condition often challenging to achieve.
@@ -155,7 +155,7 @@ Estimators are then applied to these data to compute free energy differences bet
 # Implementation
 ## Core design principles
 
-*alchemlyb* is a Python library that seeks to make doing alchemical free energy calculations easier and less error prone. 
[…]
