### 2024-08-17 1f32e88
Minor changes to the paper (#387)
* update

* remove doi

* add orcid
