+++ joss_paper/paper.bib
@@ -12,7 +12,7 @@
 
 @article{Case2005uq,
 	author = {Case, David A and Cheatham, 3rd, Thomas E and Darden, Tom and Gohlke, Holger and Luo, Ray and Merz, Jr, Kenneth M and Onufriev, Alexey and Simmerling, Carlos and Wang, Bing and Woods, Robert J},
-	journal = {J Comput Chem},
+	journal = {Journal of Computational Chemistry},
 	number = {16},
 	pages = {1668-1688},
 	title = {The {Amber} biomolecular simulation programs},
@@ -25,7 +25,7 @@
 @article{chodera2007use,
   title={Use of the Weighted Histogram Analysis Method for the Analysis of Simulated and Parallel Tempering Simulations},
   author={J. D. Chodera and Swope, W. C. and Pitera, J. W. and Seok, C. and Dill, K. A.},
-  journal={J Chem Theory Comput},
+  journal={Journal of Chemical Theory and Computation},
   volume={3},
   number={1},
   pages={26-41},
@@ -48,7 +48,7 @@
 @article{deng2009computations,
   title={Computations of standard binding free energies with molecular dynamics simulations},
   author={Deng, Y. and Roux, B.},
-  journal={J Phys Chem B},
+  journal={The Journal of Physical Chemistry B},
   volume={113},
   number={8},
   pages={2234-2246},
@@ -95,7 +95,7 @@
 @article{klimovich2015guidelines,
   title={Guidelines for the analysis of free energy calculations},
   author={Klimovich, P. V. and Shirts, M. R. and Mobley, D. L.},
-  journal={J Comput Aided Mol Des},
+  journal={Journal of Computer-Aided Molecular Design},
   volume={29},
   number={5},
   pages={397-411},
@@ -108,12 +108,13 @@
   author={Merz, Jr, K. M. and Ringe, D. and Reynolds, C. H.},
   year={2010},
   publisher={Cambridge University Press},
+  doi = {10.48550/arXiv.1309.0238},
 }
 
 @article{paliwal2011benchmark,
   title={A Benchmark Test Set for Alchemical Free Energy Transformations and Its Use to Quantify Error in Common Free Energy Methods},
   author={Paliwal, H. and Shirts, M. R.},
-  journal={J Chem Theory Comput},
+  journal={Journal of Chemical Theory and Computation},
   volume={7},
   number={12},
   pages={4115-4134},
@@ -124,7 +125,7 @@
 @article{pham2011identifying,
   title={Identifying low variance pathways for free energy calculations of molecular transformations in solution phase},
   author={Pham, T. T. and Shirts, M. R.},
-  journal={J Chem Phys},
+  journal={The Journal of Chemical Physics},
   volume={135},
   number={3},
   pages={034114},
@@ -146,7 +147,7 @@
 @article{pohorille2010good,
   title={Good practices in free-energy calculations},
   author={Pohorille, A. and Jarzynski, C. and Chipot, C.},
-  journal={J Phys Chem B},
+  journal={The Journal of Physical Chemistry B},
   volume={114},
   number={32},
   pages={10235-10253},
@@ -168,7 +169,7 @@
 @article{shirts2008statistically,
   title={Statistically optimal analysis of samples from multiple equilibrium states},
   author={Shirts, M. R. and Chodera, J. D.},
-  journal={J Chem Phys},
+  journal={The Journal of Chemical Physics},
   volume={129},
   number={12},
   pages={124105},
@@ -179,7 +180,7 @@
 @article{yang2004free,
   title={Free energy simulations: use of reverse cumulative averaging to determine the equilibrated region and the time required for convergence},
[…]
