### 2024-08-17 alchemistry/alchemlyb#386 @xiki-tempula
@srmnitc All the changes has been addressed in https://github.com/alchemistry/alchemlyb/pull/387
