+++ src/alchemlyb/parsing/lammps.py
@@ -282,11 +282,15 @@ def extract_u_nk_from_u_n(
     fep_files,
     T,
     column_lambda,
-    column_u_cross,
+    column_U,
+    column_U_cross,
     dependence=lambda x: (x),
     index=-1,
     units="real",
     prec=4,
+    ensemble="nvt",
+    pressure=None,
+    column_volume=4,
 ):
     """Produce u_nk from files containing u_n given a separable dependence on lambda.
 
@@ -299,7 +303,9 @@ def extract_u_nk_from_u_n(
         Temperature in Kelvin at which the simulation was sampled.
     columns_lambda : int
         Indices for columns (file column number minus one) representing the lambda at which the system is equilibrated
-    column_u_cross : int
+    column_U : int
+        Index for the column (file column number minus one) representing the potential energy of the system.
+    column_U_cross : int
         Index for the column (file column number minus one) representing the potential energy of the cross interactions
         between the solute and solvent.
     dependence : func, default=`lambda x : (x)`
@@ -313,6 +319,14 @@ def extract_u_nk_from_u_n(
         "real", "si"
     prec : int, default=4
         Number of decimal places defined used in ``round()`` function.
+    ensemble : str, default="nvt"
+        Ensemble from which the given data was generated. Either "nvt" or "npt" is supported where values from NVT are
+        unaltered, while those from NPT are corrected 
+    pressure : float, default=None
+        The pressure of the system in the NPT ensemble in units of energy / volume, where the units of energy and volume
+        are as recorded in the LAMMPS dump file.
+    column_volume : int, default=4
+        The column for the volume in a LAMMPS dump file. 
 
     Returns
     -------
@@ -331,6 +345,15 @@ def extract_u_nk_from_u_n(
     files = glob.glob(fep_files)
     if not files:
         raise ValueError(f"No files have been found that match: {fep_files}")
+    
+    if ensemble == "npt":
+        if pressure is None or not isinstance(pressure, float) or pressure < 0:
+            raise ValueError("In the npt ensemble, a pressure must be provided in the form of a positive float")
+    elif ensemble != "nvt":
+        raise ValueError("Only ensembles of nvt or npt are supported.")
+    else:
+        if pressure is not None:
+            raise ValueError("There is no volume correction in the nvt ensemble, the pressure value will not be used.")
 
     beta = beta_from_units(T, units)
 
@@ -338,9 +361,13 @@ def extract_u_nk_from_u_n(
         raise ValueError(
             f"Provided column for lambda must be type int. column_u_lambda: {column_lambda}, type: {type(column_lambda)}"
         )
-    if not isinstance(column_u_cross, int):
+    if not isinstance(column_U_cross, int):
         raise ValueError(
-            f"Provided column for u_cross must be type int. column_u_cross: {column_u_cross}, type: {type(column_u_cross)}"
+            f"Provided column for `U_cross` must be type int. column_U_cross: {column_U_cross}, type: {type(column_U_cross)}"
+        )
+    if not isinstance(column_U, int):
+        raise ValueError(
+            f"Provided column for `U` must be type int. column_U: {column_U}, type: {type(column_U)}"
         )
 
     lambda_values = list(
@@ -350,25 +377,31 @@ def extract_u_nk_from_u_n(
 
     u_nk = pd.DataFrame(columns=["time", "fep-lambda"] + lambda_values)
     lc = len(lambda_values)
-    col_indices = [0, column_lambda, column_u_cross]
+    col_indices = [0, column_lambda, column_U, column_U_cross]
+    if ensemble == "npt":
+        col_indices.append(column_volume)
 
     for file in files:
         if not os.path.isfile(file):
[…]
