### 2024-08-23 fa5a539
Bug fix, inputs for extract_dHdl
