+++ setup.py
@@ -29,7 +29,6 @@ setup(
         "Operating System :: Microsoft :: Windows ",
         "Programming Language :: Python",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
         "Programming Language :: Python :: 3.11",
         "Programming Language :: Python :: 3.12",
@@ -43,7 +42,7 @@ setup(
     license="BSD",
     long_description=open("README.md").read(),
     long_description_content_type="text/markdown",
-    python_requires=">=3.9",
+    python_requires=">=3.10",
     tests_require=["pytest", "alchemtest"],
     install_requires=[
         "numpy",
