### 2024-08-27 5f9fff7
Refactor moving_average to align with forward_backward_convergence function
