+++ src/alchemlyb/convergence/convergence.py
@@ -30,7 +30,8 @@ def forward_backward_convergence(
     Parameters
     ----------
     df_list : list
-        List of DataFrame of either dHdl or u_nk.
+        List of DataFrame of either dHdl or u_nk, where each represents a
+        different value of lambda.
     estimator : {'MBAR', 'BAR', 'TI'}
         Name of the estimators.
         See the important note below on the use of "MBAR".
@@ -94,7 +95,18 @@ def forward_backward_convergence(
         # select estimator class by name
         my_estimator = estimators_dispatch[estimator](**kwargs)
         logger.info(f"Use {estimator} estimator for convergence analysis.")
-
+    
+    # Check that each df in the list has only one value of lambda
+    for i, df in enumerate(df_list):
+        lambda_values = np.unique([x[1] for x in df.index.to_numpy()])
+        if len(lambda_values) > 1:
+            if len(df_list[0].index[0]) > 2:
+                lambda_values = np.unique([x[2] for x in df.index.to_numpy()])
+                if len(lambda_values) > 1:
+                    raise ValueError("Provided DataFrame, df_list[{}] has more than one lambda value".format(i))
+            else:
+                raise ValueError("Provided DataFrame, df_list[{}] has more than one lambda value".format(i))
+            
     logger.info("Begin forward analysis")
     forward_list = []
     forward_error_list = []
@@ -396,14 +408,15 @@ def moving_average(df_list, estimator="MBAR", num=10, **kwargs):
 
     Generate the free energy estimate for a series of blocks in time,
     with the specified number of equally spaced points.
-    For example, setting `num` to 10 would give the forward
-    convergence which is the free energy estimate from the first 10%, then the
+    For example, setting `num` to 10 would give the moving average
+    which is the free energy estimate from the first 10% alone, then the
     next 10% ... of the data.
 
     Parameters
     ----------
     df_list : list
-        List of DataFrame of either dHdl or u_nk.
+        List of DataFrame of either dHdl or u_nk, where each represents a
+        different value of lambda.
     estimator : {'MBAR', 'BAR', 'TI'}
         Name of the estimators.
         See the important note below on the use of "MBAR".
@@ -444,72 +457,44 @@ def moving_average(df_list, estimator="MBAR", num=10, **kwargs):
         estimator_fit = estimators_dispatch[estimator](**kwargs).fit
         logger.info(f"Use {estimator} estimator for convergence analysis.")
 
-    logger.info("Check indices")
-    if estimator in ["MBAR"]:
-        index_1 = [
-            np.unique(np.array([x[1] for x in data.index.to_numpy()]))
-            for data in df_list
-        ]
-        if len(np.unique(index_1)) == 1 and len(df_list[0].index[0]) > 2:
-            index_2 = [
-                np.unique(np.array([x[2] for x in data.index.to_numpy()]))
-                for data in df_list
-            ]
-            if len(np.unique(index_2)) > 1:
-                raise ValueError(
-                    "Restrict to a single fep-lambda value for a meaningful result. index[2] for each file"
-                    " in df_list: {}".format(index_2)
-                )
-        elif len(np.unique(index_1)) != 1:
-            raise ValueError(
-                "Restrict to a single fep-lambda value for a meaningful result. index[1] for each file"
-                " in df_list: {}".format(index_1)
-            )
-    elif estimator in ["BAR"]:
-        index_1 = [
-            np.unique(np.array([x[1] for x in data.index.to_numpy()]))
-            for data in df_list
-        ]
-        if len(np.unique(index_1)) == 1 and len(df_list[0].index[0]) > 2:
-            index_2 = [
[…]
