### 2024-08-28 75aa16e
Update test_convergence and lambda tests in convergence.moving_average
