### 2024-08-28 ca391c3
Adjust convergence.py and tests for codecoverage
