### 2024-08-30 8da4b99
Add LAMMPS unit conversion for pV term in npt u_r
