### 2024-08-30 2d5fe48
Fixed kT multiplication issue in dHdl
