### 2024-08-30 ce88763
Update moving_average to block_average for more accurate descriptive name
