+++ src/alchemlyb/convergence/__init__.py
@@ -1,1 +1,1 @@
-from .convergence import forward_backward_convergence, fwdrev_cumavg_Rc, A_c, moving_average
+from .convergence import forward_backward_convergence, fwdrev_cumavg_Rc, A_c, block_average
