### 2024-09-03 bd2c23a
Resolve deprecation warnings
