### 2024-09-03 afec5e2
Update pandas syntax to avoid FutureWarnings
