### 2024-09-03 c286a6a
Update test to align with changed handling of dfs of different length in block_average
