### 2024-09-05 #6934 @orbeckst
Final check list

   **release [2.3.2](https://github.com/alchemistry/alchemlyb/releases/tag/2.3.2)**

  zenodo did not automatically archive the 2.3.2 release — need to investigate

  (The 2.3.1 release is https://doi.org/10.5281/zenodo.12692737 .)

  Do you want the version-specific DOI or the generic DOI [10.5281/zenodo.596103](https://zenodo.org/doi/10.5281/zenodo.596103) (which points to the latest released version)?

  The author list in the zenodo archive is auto-generated by zenodo. I don't know how we could change it.

  Similarly, we do not store ORCIDs anywhere so I don't know how we would be able to add them to the zenodo archive. Do you suggest that we have to include them in the AUTHORS file??

  The LICENSE file with the BSD-3 clause license is always included in the archive.
