### 2024-09-05 alchemistry/alchemlyb#393 @orbeckst
zenodo did not create archive for 2.3.2
The [2.3.2 release](https://github.com/alchemistry/alchemlyb/releases/tag/2.3.2) was cut 2 weeks ago but the [latest release on zenodo](https://zenodo.org/doi/10.5281/zenodo.596103) is still 2.3.1 https://zenodo.org/records/12692737 .

This is a problem because for the JOSS paper we need an archive of 2.3.2 (or later) https://github.com/openjournals/joss-reviews/issues/6934#issuecomment-2332700243 .
