### 2024-09-05 alchemistry/alchemlyb#393 @orbeckst
Maybe there are currently issues with zenodo?

We can also try making a 2.3.3 and see if this triggers zenodo.
