### 2024-09-05 alchemistry/alchemlyb#393 @orbeckst
I got a [409](https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/409) response ("Conflict") from zenodo when I manually tried to update the release using the script from https://github.com/zenodo/zenodo/issues/1463#issuecomment-1007602828 . Maybe just try a 2.3.3 release.

```python
#!/usr/bin/env python
# (c) Richard Shaw 2022 jrs65 https://github.com/jrs65
# https://github.com/zenodo/zenodo/issues/1463#issuecomment-1007602828

import requests

# Fill these in...
repo = "alchemistry/alchemlyb"
token = "<token from settings webhook>"

headers = {"Accept": "application/vnd.github.v3+json"}

repo_response = requests.get(f"https://api.github.com/repos/{repo}", headers=headers)
release_response = requests.get(f"https://api.github.com/repos/{repo}/releases", headers=headers)

latest_release = release_response.json()[0]

payload = {"action": "published", "release": latest_release, "repository": repo_response.json()}

print(f"Latest release: {latest_release['name']}")
#print(payload)

submit_response = requests.post(
    f"https://zenodo.org/api/hooks/receivers/github/events/?access_token={token}",
    json=payload
)

print(submit_response)

```
