### 2024-09-05 alchemistry/alchemlyb#393 @orbeckst
I'll try a webhook retrigger following https://github.com/zenodo/zenodo/issues/1181#issuecomment-1015671213 ...
