### 2024-09-06 alchemistry/alchemlyb#393 @orbeckst
For reasons that I don't understand, the zenodo switch was "off" for alchemistry/alchemlyb. I now set it to "on".

This *may* already have created a new release but I didn't check right away and also tried the manual approach again:

I repeated the push with the script and got
```
Latest release: 2.3.2
<Response [202]>
```
where [202](https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/202) is "Accepted".

We now have the 2.3.2 archive on zenodo https://zenodo.org/records/13699232 !
