### 2024-09-06 alchemistry/alchemlyb#393 @orbeckst
We now have two hooks:
* old: "never triggered" https://zenodo.org/api/hooks/receivers/github/events/?access_token=u...6
* new: "last delivery: 2024-09-05 16:57:19" https://zenodo.org/api/hooks/receivers/github/events/?access_token=r...u

I deleted the first, "never used" hook.
