### 2024-09-09 #6934 @orbeckst
@srmnitc I addressed the outstanding requirements as far as I could. I have a question about the authorship requirements for the zenodo package (see https://github.com/openjournals/joss-reviews/issues/6934#issuecomment-2332700243 above without check mark). Could you please clarify what's required? Thanks.
