### 2024-09-10 f51d390
Remove incorrect popagation of error in BAR
