+++ src/alchemlyb/convergence/convergence.py
@@ -181,14 +181,9 @@ def _forward_backward_convergence_estimate(
         my_estimator.initial_f_k = result.delta_f_.iloc[0, :]
     mean = result.delta_f_.iloc[0, -1]
     if estimator.lower() == "bar":
-        error = np.sqrt(
-            sum(
-                [
-                    result.d_delta_f_.iloc[i, i + 1] ** 2
-                    for i in range(len(result.d_delta_f_) - 1)
-                ]
-            )
-        )
+        # See https://github.com/alchemistry/alchemlyb/pull/60#issuecomment-430720742
+        # Error estimate generated by BAR ARE correlated
+        error = np.nan
     else:
         error = result.d_delta_f_.iloc[0, -1]
     if estimator.lower() == "mbar" and error > error_tol:
@@ -486,15 +481,9 @@ def block_average(df_list, estimator="MBAR", num=10, **kwargs):
 
         average_list.append(result.delta_f_.iloc[0, -1])
         if estimator.lower() == "bar":
-            error = np.sqrt(
-                sum(
-                    [
-                        result.d_delta_f_.iloc[i, i + 1] ** 2
-                        for i in range(len(result.d_delta_f_) - 1)
-                    ]
-                )
-            )
-            average_error_list.append(error)
+            # See https://github.com/alchemistry/alchemlyb/pull/60#issuecomment-430720742
+            # Error estimate generated by BAR ARE correlated
+            average_error_list.append(np.nan)
         else:
             average_error_list.append(result.d_delta_f_.iloc[0, -1])
         logger.info(
