### 2024-09-10 a179048
Add tests and error catch for ill constructed BAR input, u_nk
