+++ src/alchemlyb/estimators/mbar_.py
@@ -79,6 +79,11 @@ class MBAR(BaseEstimator, _EstimatorMixOut):
         default value for `method` was changed from "hybr" to "robust"
     .. versionchanged:: 2.1.0
         `n_bootstraps` option added.
+    .. versionchanged:: 2.4.0
+       Handle initial estimate, initial_f_k, from bar in the instance
+       that not all lambda states represented as column headers are 
+       represented in the indices of u_nk.
+    
     """
 
     def __init__(
@@ -135,11 +140,6 @@ class MBAR(BaseEstimator, _EstimatorMixOut):
             )
             bar.fit(u_nk)
             initial_f_k = bar.delta_f_.iloc[0, :]
-            states = [
-                x
-                for i, x in enumerate(self._states_[:-1])
-                if N_k[i] > 0 and N_k[i + 1] > 0
-            ]
             if len(bar.delta_f_.iloc[0, :]) != len(self._states_):
                 states = [
                     x
