### 2024-09-11 #6934 @srmnitc
@orbeckst One can tune the metadata in Zenodo using a CIFF file: https://citation-file-format.github.io/. You can see an example for one of my packages [here](https://github.com/pyscal/atomRDF/blob/main/CITATION.cff). You would need to make another release after, probably with updated version. There is also contributer role available, so if you want to credit other people who are not in the paper, that could be done. Would this be okay for you?
