### 2024-09-11 alchemistry/alchemlyb#394 @orbeckst
create CFF citation file
Add a [citation CIFF file](https://citation-file-format.github.io/) to the repo so that zenodo picks up all authors. We need this for the final JOSS paper acceptance https://github.com/openjournals/joss-reviews/issues/6934#issuecomment-2342882680.
- [x] all authors and contributors
- [x] add the JOSS paper #328

After adding the CIFF file we need to make another patch level release.

See https://github.com/openjournals/joss-reviews/issues/6934#issuecomment-2342882680 for background:

> One can tune the metadata in Zenodo using a CIFF file: https://citation-file-format.github.io/. You can see an example for one of my packages [here](https://github.com/pyscal/atomRDF/blob/main/CITATION.cff). You would need to make another release after, probably with updated version. There is also contributer role available, so if you want to credit other people who are not in the paper, that could be done.
