### 2024-09-11 b5ffe13
initial CITATION.cff file
- generated with https://bit.ly/cffinit
- authors and order from AUTHORS
- affiliations and ORCID from JOSS paper
- emails from paper or individuals

Co-authored-by: David L. Dotson <dotsdl@gmail.com>
