### 2024-09-11 50bbbab
add JOSS paper as preferred citation to CITATION
- inferred DOI from proof
- assume final acceptance in 2024...
