+++ CITATION.cff
@@ -69,7 +69,7 @@ authors:
     family-names: Alibay
     orcid: 'https://orcid.org/0000-0001-5787-9130'
     affiliation: Open Molecular Software Foundation
-  - given-names: Pascal
+  - given-names: Pascal T.
     family-names: Merz
     affiliation: PM Scientific Consulting
     orcid: 'https://orcid.org/0000-0002-7045-8725'
@@ -111,3 +111,85 @@ keywords:
   - molecular dynamics
   - alchemistry
 license: BSD-3-Clause
+preferred-citation:
+  title: 'alchemlyb: the simple alchemistry library'
+  authors:
+    - given-names: Zhiyi
+      family-names: Wu
+      affiliation: Exscientia plc
+      orcid: 'https://orcid.org/0000-0002-7615-7851'
+      email: william@zhiyiwu.me  
+    - given-names: David L.
+      family-names: Dotson
+      orcid: 'https://orcid.org/0000-0001-5879-2942'
+      affiliation: Arizona State University
+    - given-names: Irfan
+      family-names: Alibay
+      orcid: 'https://orcid.org/0000-0001-5787-9130'
+      affiliation: Open Molecular Software Foundation
+    - given-names: Bryce
+      family-names: Allen
+      orcid: 'https://orcid.org/0000-0002-0804-8127'
+      affiliation: Differentiated Therapeutics
+    - given-names: Mohammad Soroush
+      family-names: Barhaghi
+      orcid: 'https://orcid.org/0000-0001-8226-7347'
+      affiliation: Wayne State University
+    - given-names: Jérôme
+      family-names: Hénin
+      orcid: 'https://orcid.org/0000-0003-2540-4098'
+      affiliation: CNRS
+    - given-names: Thomas T.
+      family-names: Joseph
+      orcid: 'https://orcid.org/0000-0003-1323-3244'
+      affiliation: University of Pennsylvania      
+    - given-names: Ian M.
+      family-names: Kenney
+      orcid: 'https://orcid.org/0000-0002-9749-8866'
+      affiliation: Arizona State University
+    - given-names: Hyungro
+      family-names: Lee
+      orcid: 'https://orcid.org/0000-0002-4221-7094'
+      affiliation: Pacific Northwest National Laboratory
+    - given-names: Haoxi
+      family-names: Li
+      orcid: 'https://orcid.org/0009-0004-8369-1042'
+      affiliation: University of North Carolina      
+    - given-names: Victoria
+      family-names: Lim
+      orcid: 'https://orcid.org/0000-0003-4030-9312'
+      affiliation: University of California Irvine      
+    - given-names: Shuai
+      family-names: Liu
+      orcid: 'https://orcid.org/0000-0002-8632-633X'
+      affiliation: Silicon Therapeutics LLC
+    - given-names: Domenico
+      family-names: Marson
+      orcid: 'https://orcid.org/0000-0003-1839-9868'
+      affiliation: University of Trieste
+    - given-names: Pascal T.
+      family-names: Merz
+      affiliation: PM Scientific Consulting
+      orcid: 'https://orcid.org/0000-0002-7045-8725'      
+    - given-names: Alexander
+      family-names: Schlaich
+      orcid: 'https://orcid.org/0000-0002-4250-363X'
+      affiliation: University of Stuttgart
+    - given-name: David
+      family-name: Mobley
+      orcid: 'https://orcid.org/0000-0002-1083-5533'
+      affiliation: University of California Irvine
+    - given-names: Michael R.
+      family-names: Shirts
+      orcid: 'https://orcid.org/0000-0003-3249-1097'
+      affiliation: University of Colorado Boulder
+    - given-names: Oliver
[…]
