### 2024-09-11 d12d662
added references for MBAR and equilibration detection
