### 2024-09-11 ba25618
inserted additional JOSS paper authors in software author list
- added David Mobley @davidlmobley (original conception and design of alchemlyb)
- added Michael Shirts @mrshirts (original conception and design of alchemlyb)
