### 2024-09-12 8dcd2d4
add authors to CITATION who had NOT been in AUTHORS
Added all authors that were mentioned in the Acknowledgements of the JOSS
paper but were not in AUTHORS

- Wei-Tse Hsu @wehs7661 (2020) for code clean-up in a732380cc6b9571eca7eb694e3acd8adeda4ed78
- Jan Janssen  @jan-janssen (2022) for creating the conda-forge package
- Shujie Fan @VOD555 (2022) for initial code for fractional equilibration time
- Helmut Carter @helmutcarter (2024) for doc fix in #356
