### 2024-09-12 #6934 @orbeckst
@srmnitc I am adding a CITATION.cff file in https://github.com/alchemistry/alchemlyb/pull/395 . Everyone from the AUTHORS file and all paper authors are in the top level `authors` entry as well as anyone whom we listed in Acknowledgements.

> There is also contributer role available, so if you want to credit other people who are not in the paper, that could be done. Would this be okay for you?

I haven't found a contributor role in the [CFF schema guide](https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md). Please let me know if you know more. However, everyone who is an author of the paper or is mentioned in Acknowledgements is included in the author list of the CITATION file so perhaps it's not necessary to distinguish so finely.

I added a preferred-citation entry for the JOSS paper (assuming that it will get accepted) and this entry only contains the authors of the paper.
