+++ docs/visualisation/alchemlyb.visualisation.plot_block_average.rst
@@ -1,0 +1,6 @@
+﻿alchemlyb.visualisation.plot\_block\_average
+=============================================
+
+.. currentmodule:: alchemlyb.visualisation
+
+.. autofunction:: plot_block_average
