+++ src/alchemlyb/visualisation/__init__.py
@@ -1,4 +1,4 @@
-from .convergence import plot_convergence
+from .convergence import plot_convergence, plot_block_average
 from .dF_state import plot_dF_state
 from .mbar_matrix import plot_mbar_overlap_matrix
 from .ti_dhdl import plot_ti_dhdl
