### 2024-09-16 d97d682
add @jaclark5 to AUTHORS and CITATION.cff
AUTHORS entry had been forgotten in PR #381
