+++ CITATION.cff
@@ -105,6 +105,10 @@ authors:
     family-names: Shirts
     orcid: 'https://orcid.org/0000-0003-3249-1097'
     affiliation: University of Colorado Boulder
+  - given-names: Jennifer A.
+    family-names: Clark
+    orcid: https://orcid.org/0000-0003-4897-5651
+    affiliation: National Institute of Standards and Technology
 identifiers:
   - type: swh
     value: 'swh:1:dir:4cd266754be413312ee7ef3f1f7fbe05b38192cf'
