### 2024-09-17 aaf5218
Resolve bug in bar_model.delta_f_ creation (#397)
* fix bug introduced in PR #381: there was a change to creating the delta_f_ matrix,
  which resulted in the columns and indices being tuples that were in the wrong order 
  for single lambda computations.
* ensure that columns are in the correct order by explicitly sorting
* add a test for the delta_f_ columns
