+++ src/alchemlyb/visualisation/convergence.py
@@ -59,7 +59,7 @@ def plot_convergence(dataframe, units=None, final_error=None, ax=None):
      .. versionchanged:: 0.6.0
          data now takes in dataframe
      .. versionadded:: 0.4.0
-     
+
     """
     if units is not None:
         dataframe = get_unit_converter(units)(dataframe)
@@ -192,7 +192,7 @@ def plot_block_average(dataframe, units=None, final_error=None, ax=None):
      `Alchemical Analysis <https://github.com/MobleyLab/alchemical-analysis>`_.
 
      .. versionadded:: 2.4.0
-     
+
     """
     if units is not None:
         dataframe = get_unit_converter(units)(dataframe)
