+++ src/alchemlyb/parsing/util.py
@@ -1,6 +1,7 @@
 """Collection of utilities used by many parsers.
 
 """
+
 import bz2
 import gzip
 import os
