+++ src/alchemlyb/parsing/gmx.py
@@ -1,6 +1,7 @@
 """Parsers for extracting alchemical data from `Gromacs <http://www.gromacs.org/>`_ output files.
 
 """
+
 import numpy as np
 import pandas as pd
 
