### 2024-09-17 a3b2c08
replace setup.py,cfg with pyproject.toml
- fix #385
- use pyproject.toml instead of setup.py (note: may need to change README for PyPi to exclude
  banners)
- remove versioneer and use versioningit (use alchemlyb.__version__ directly where
  the version is needed, e.g., for sphinx docs)
- updated CHANGES for 2.4.0
