+++ src/alchemlyb/__init__.py
@@ -2,10 +2,7 @@ from functools import wraps
 
 import pandas as pd
 
-from ._version import get_versions
-
-__version__ = get_versions()["version"]
-del get_versions
+from ._version import __version__
 
 
 def pass_attrs(func):
