### 2024-09-17 #6934 @orbeckst
@srmnitc we now have a 2.4.0 release with a nicely-formatted zenodo archive https://zenodo.org/records/13774657
* release **2.4.0** (on [GitHub](https://github.com/alchemistry/alchemlyb/releases/tag/2.4.0) and [PyPi](https://pypi.org/project/alchemlyb/), conda-forge package should come through soon)
* doi https://doi.org/10.5281/zenodo.13774657
* authors included in [zenodo release 2.4.0](https://zenodo.org/records/13774657) with ORCIDs
