### 2024-09-18 #6934 @srmnitc
@orbeckst The title on the paper, and on Zenodo should be the same, and this would also need to be changed. I am sorry that I did not spot this before! Really sorry about that. Could you please change the title on Zenodo. I believe this doesnt need a rerelease. You should be able to edit this directly on Zenodo.
