### 2024-09-18 alchemistry/alchemlyb#402 @orbeckst
make zenodo title agree with paper title
For the JOSS paper https://github.com/openjournals/joss-reviews/issues/6934#issuecomment-2358113559 (issue #71 , PR #328 ) we are asked to make the title of the [zenodo archive](https://zenodo.org/records/13774657) (currently "alchemlyb") be the same as the paper (PR #328,  title: 'alchemlyb: the simple alchemistry library').

This will require
- [ ] update CITATION file
- [ ] make a new release 2.4.1
