### 2024-09-18 #6934 @orbeckst
I didn’t find a way to change the title in zenodo so we will do a 2.4.1 release with updated title in the CITATION file.
