### 2024-09-18 8b5820b
Add enthalpy and entropy calculations
