+++ src/alchemlyb/estimators/bar_.py
@@ -1,5 +1,6 @@
 import numpy as np
 import pandas as pd
+import pymbar
 from pymbar.other_estimators import bar as BAR_
 from sklearn.base import BaseEstimator
 
@@ -35,6 +36,20 @@ class BAR(BaseEstimator, _EstimatorMixOut):
         The estimated statistical uncertainty (one standard deviation) in
         dimensionless free energy differences.
 
+    delta_h_ : DataFrame
+        The estimated dimensionless enthalpy difference between each state.
+
+    d_delta_h_ : DataFrame
+        The estimated statistical uncertainty (one standard deviation) in
+        dimensionless enthalpy differences.
+
+    delta_s_ : DataFrame, optional
+        The estimated dimensionless entropy difference between each state.
+
+    d_delta_s_ : DataFrame
+        The estimated statistical uncertainty (one standard deviation) in
+        dimensionless entropy differences.
+
     states_ : list
         Lambda states for which free energy differences were obtained.
 
@@ -57,6 +72,7 @@ class BAR(BaseEstimator, _EstimatorMixOut):
     .. versionchanged:: 2.4.0
        Added assessment of lambda states represented in the indices of u_nk
        to provide meaningful errors to ensure proper use.
+       Added computation of enthalpy and entropy with 2-state MBAR
 
     """
 
@@ -76,10 +92,12 @@ class BAR(BaseEstimator, _EstimatorMixOut):
         # handle for pymbar.BAR object
         self._bar = None
 
-    def fit(self, u_nk):
+    def fit(self, u_nk, use_mbar=False, compute_entropy_enthalpy=False):
         """
         Compute overlap matrix of reduced potentials using
         Bennett acceptance ratio.
+
+        Sets the attributes: delta_f_ and d_delta_f_
 
         Parameters
         ----------
@@ -87,7 +105,21 @@ class BAR(BaseEstimator, _EstimatorMixOut):
             u_nk[n,k] is the reduced potential energy of uncorrelated
             configuration n evaluated at state k.
 
+        use_mbar : bool, optional, default=False
+            Use 2-state MBAR instead of BAR. This will allow for the
+            calculation of enthalpic and entropic contributions.
+
+        compute_entropy_enthalpy : bool, optional, default=False
+            Compute entropy and enthalpy from 2-state MBAR. Note
+            that ``use_mbar`` must be ``True``.
+
         """
+
+        if compute_entropy_enthalpy and not use_mbar:
+            raise ValueError(
+                "Cannot compute the enthalpy and entropy with BAR, set use_mbar=True."
+            )
+
         # sort by state so that rows from same state are in contiguous blocks
         u_nk = u_nk.sort_index(level=u_nk.index.names[1:])
 
@@ -114,6 +146,11 @@ class BAR(BaseEstimator, _EstimatorMixOut):
         # Now get free energy differences and their uncertainties for each step
         deltas = np.array([])
         d_deltas = np.array([])
+        if compute_entropy_enthalpy:
+            deltas_h = np.array([])
+            d_deltas_h = np.array([])
+            deltas_s = np.array([])
+            d_deltas_s = np.array([])
         for k in range(len(N_k) - 1):
             if N_k[k] == 0 or N_k[k + 1] == 0:
                 continue
@@ -137,10 +174,38 @@ class BAR(BaseEstimator, _EstimatorMixOut):
                 relative_tolerance=self.relative_tolerance,
                 verbose=self.verbose,
             )
+            if use_mbar:  # now determine df and ddf using pymbar.MBAR
+                tmp_u_nk = u_nk.iloc[
[…]
