### 2024-09-19 263a431
make repo title identical to JOSS paper title (#404)
- close #402
- use same title in CITATION.cff for paper and project
  (required for the JOSS paper publication #71)
- added DOI for preferred citation in note because that is the
  only hint that shows up on zenodo)
