+++ CITATION.cff
@@ -3,10 +3,11 @@
 # Visit https://bit.ly/cffinit to generate yours today!
 
 cff-version: 1.2.0
-title: alchemlyb
+title: 'alchemlyb: the simple alchemistry library'
 message: >-
   If you use this software, please cite it using the
-  preferred citation together with any other references.
+  preferred citation (JOSS DOI 10.21105/joss.06934) together
+  with any other references.
 type: software
 authors:
   - email: david@datryllic.com
