### 2024-09-19 ac29639
[doc] tutorial: use alchemlyb.concat (#399)
replace `pandas.concat()` with `alchemlyb.concat()` in the tutorial (given that we explicitly tell users to use it)
