### 2024-09-19 37d6a17
Add enthalpy and entropy calculations
