+++ src/alchemlyb/estimators/base.py
@@ -5,6 +5,10 @@ class _EstimatorMixOut:
     _d_delta_f_ = None
     _delta_f_ = None
     _states_ = None
+    _d_delta_h_ = None
+    _delta_h_ = None
+    _d_delta_s_ = None
+    _delta_s_ = None
 
     @property
     def d_delta_f_(self):
@@ -13,6 +17,22 @@ class _EstimatorMixOut:
     @property
     def delta_f_(self):
         return self._delta_f_
+
+    @property
+    def d_delta_h_(self):
+        return self._d_delta_h_
+
+    @property
+    def delta_h_(self):
+        return self._delta_h_
+
+    @property
+    def d_delta_s_(self):
+        return self._d_delta_s_
+
+    @property
+    def delta_s_(self):
+        return self._delta_s_
 
     @property
     def states_(self):
