### 2024-09-19 214bbfe
Add bootstrapping to BAR().fit(use_mbar=True) and adapt tests for code coverage
