### 2024-09-19 5574c3b
Update get_group syntax
