### 2024-09-19 187246e
avoid pandas futurewarning for grouping with get_group()  (#407)
* fix #407
* Update pandas syntax to avoid FutureWarnings:
   FutureWarning: When grouping with a length-1 list-like, you will need to pass a length-1 tuple to get_group in a future version of pandas. Pass `(name,)` instead of `name` to silence this warning.
* Update CHANGES
