### 2024-09-19 5e1a393
Update method default in bar when use_mbar==True
