### 2024-09-19 4f7e2db
add @jaclark5 to acknowledgements
- acknowledging recent additions to 2.4.0 and 2.4.1 release (after the review of the manuscript)
- approved in https://github.com/alchemistry/alchemlyb/pull/328#discussion_r1767590941
