### 2024-09-19 #6934 @orbeckst
@srmnitc we have released alchemlyb [2.4.1](https://github.com/alchemistry/alchemlyb/releases/tag/2.4.1) with a Zenodo archive with DOI [10.5281/zenodo.13799342](https://doi.org/10.5281/zenodo.13799342).

Please note that we added one author to the acknowledgements section of the manuscript (Jennifer A. Clark, with her permission) in https://github.com/alchemistry/alchemlyb/pull/328/commits/4f7e2dbf024a2d552ced51acbe0fd4153ea1e803 because she contributed code to the 2.4.0 and 2.4.1 releases that were completed _after_ the submission and review of the paper. In this way, all authors visible on the zotero page are also mentioned in the paper (either as authors or in Acknowledgements).

I believe we have now completed all requirements.
