+++ src/alchemlyb/parsing/lammps.py
@@ -86,6 +86,7 @@ def beta_from_units(T, units):
 
     return beta
 
+
 def energy_from_units(units):
     """Output conversion factor for pressure * volume to LAMMPS energy units
 
@@ -109,8 +110,8 @@ def energy_from_units(units):
     .. versionadded:: 2.4.1
 
     """
-    if units == "real":  # E in kcal/mol, Vol in Å^3, pressure in atm 
-        beta = constants.atm * constants.angstrom**3 / 1e+3 * kJ2kcal * constants.N_A
+    if units == "real":  # E in kcal/mol, Vol in Å^3, pressure in atm
+        beta = constants.atm * constants.angstrom**3 / 1e3 * kJ2kcal * constants.N_A
     elif units == "lj":  # Nondimensional E scaled by epsilon
         beta = 1
     elif units == "metal":  # E in eV, vol in Å^3, pressure in bar
@@ -121,11 +122,15 @@ def energy_from_units(units):
         beta = 1
     elif units == "electron":  # E in Hartrees, vol in Bohr^3, pressure in Pa
         Hartree2J = 4.3597447222060e-8
-        Bohr2m = 5.29177210544e+11
+        Bohr2m = 5.29177210544e11
         beta = 1 / Hartree2J / Bohr2m**3
-    elif units == "micro":  # E in picogram-micrometer^2/microsecond^2, vol in um^3, pressure in picogram/(micrometer-microsecond^2)
+    elif (
+        units == "micro"
+    ):  # E in picogram-micrometer^2/microsecond^2, vol in um^3, pressure in picogram/(micrometer-microsecond^2)
         beta = 1
-    elif units == "nano":  # E in attogram-nanometer^2/nanosecond^2, vol in nm^3, pressure in attogram/(nanometer-nanosecond^2)
+    elif (
+        units == "nano"
+    ):  # E in attogram-nanometer^2/nanosecond^2, vol in nm^3, pressure in attogram/(nanometer-nanosecond^2)
         beta = 1
     else:
         raise ValueError(
@@ -374,12 +379,12 @@ def extract_u_nk_from_u_n(
         Number of decimal places defined used in ``round()`` function.
     ensemble : str, default="nvt"
         Ensemble from which the given data was generated. Either "nvt" or "npt" is supported where values from NVT are
-        unaltered, while those from NPT are corrected 
+        unaltered, while those from NPT are corrected
     pressure : float, default=None
         The pressure of the system in the NPT ensemble in units of energy / volume, where the units of energy and volume
         are as recorded in the LAMMPS dump file.
     column_volume : int, default=4
-        The column for the volume in a LAMMPS dump file. 
+        The column for the volume in a LAMMPS dump file.
 
     Returns
     -------
@@ -399,15 +404,19 @@ def extract_u_nk_from_u_n(
     files = glob.glob(fep_files)
     if not files:
         raise ValueError(f"No files have been found that match: {fep_files}")
-    
+
     if ensemble == "npt":
         if pressure is None or not isinstance(pressure, float) or pressure < 0:
-            raise ValueError("In the npt ensemble, a pressure must be provided in the form of a positive float")
+            raise ValueError(
+                "In the npt ensemble, a pressure must be provided in the form of a positive float"
+            )
     elif ensemble != "nvt":
         raise ValueError("Only ensembles of nvt or npt are supported.")
     else:
         if pressure is not None:
-            raise ValueError("There is no volume correction in the nvt ensemble, the pressure value will not be used.")
+            raise ValueError(
+                "There is no volume correction in the nvt ensemble, the pressure value will not be used."
+            )
 
     beta = beta_from_units(T, units)
 
[…]
