### 2024-09-23 #6934 @Kevin-Mattheus-Moerman
@orbeckst as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I will now process these final checks:

- [x] Review completed, e.g. all review tick boxes ticked.
- [x] Software license tag listed here matches a tagged release

- [ ] For the paper affiliations please spell out USA as United States of America
- [ ] The archive listed author set or order does not match the paper author set.
