### 2024-09-24 27caba8
paper affiliations: spell out USA
address https://github.com/openjournals/joss-reviews/issues/6934#issuecomment-2367669802
