### 2024-09-24 #6934 @orbeckst
@Kevin-Mattheus-Moerman I spelled out USA in the affiliations.

However, I do not know how to address the point about the authorship list in the zenodo archive. The author list of the paper differs from the list of contributors on the software. Not all software authors chose to be authors on the paper or their contributions did not warrant authorship (as already discussed with @srmnitc ). However, every author on the zenodo archive is either an author of the paper or mentioned in the Acknowledgements of the paper. I cannot change the author list on the zenodo archive to conform to the paper list.

If JOSS insists on these two lists being identical I will have to manually create a zenodo submission with a release tar ball included. (I was happy that I could use the automatic zenodo archive, that is tightly linked to the GitHub repo, for the JOSS paper submission but if needed will create a manual dump archive.)
