### 2024-09-24 98fd197
Added commenting extract_u_nk
