+++ src/alchemlyb/parsing/lammps.py
@@ -112,7 +112,9 @@ def energy_from_units(units):
     """
     if units == "real":  # E in kcal/mol, Vol in Å^3, pressure in atm
         beta = constants.atm * constants.angstrom**3 / 1e3 * kJ2kcal * constants.N_A
-    elif units == "lj":  # Nondimensional E scaled by epsilon
+    elif (
+        units == "lj"
+    ):  # Nondimensional E scaled by epsilon, vol in sigma^3, pressure in epsilon / sigma^3
         beta = 1
     elif units == "metal":  # E in eV, vol in Å^3, pressure in bar
         beta = constants.bar * constants.angstrom**3 / constants.eV
@@ -124,13 +126,11 @@ def energy_from_units(units):
         Hartree2J = 4.3597447222060e-8
         Bohr2m = 5.29177210544e11
         beta = 1 / Hartree2J / Bohr2m**3
-    elif (
-        units == "micro"
-    ):  # E in picogram-micrometer^2/microsecond^2, vol in um^3, pressure in picogram/(micrometer-microsecond^2)
+    elif units == "micro":
+        # E in picogram-micrometer^2/microsecond^2, vol in um^3, pressure in picogram/(micrometer-microsecond^2)
         beta = 1
-    elif (
-        units == "nano"
-    ):  # E in attogram-nanometer^2/nanosecond^2, vol in nm^3, pressure in attogram/(nanometer-nanosecond^2)
+    elif units == "nano":
+        # E in attogram-nanometer^2/nanosecond^2, vol in nm^3, pressure in attogram/(nanometer-nanosecond^2)
         beta = 1
     else:
         raise ValueError(
@@ -441,23 +441,24 @@ def extract_u_nk_from_u_n(
     u_nk = pd.DataFrame(columns=["time", "fep-lambda"] + lambda_values)
     lc = len(lambda_values)
     col_indices = [0, column_lambda, column_U, column_U_cross]
+    columns = ["time", "fep-lambda", "U", "U_cross"]
     if ensemble == "npt":
         col_indices.append(column_volume)
+        columns.append("volume")
 
     for file in files:
         if not os.path.isfile(file):
             raise ValueError("File not found: {}".format(file))
 
         tmp_data = pd.read_csv(file, sep=" ", comment="#", header=None)
-        lx = len(tmp_data.columns)
-        if [False for x in col_indices if x >= lx]:
+        ind = [x for x in col_indices if x > len(tmp_data.columns)]
+        if len(ind) > 0:
             raise ValueError(
-                "Number of columns, {}, is less than index: {}".format(lx, col_indices)
+                "Number of columns, {}, is less than indices: {}".format(
+                    len(tmp_data.columns), ind
+                )
             )
         data = tmp_data.iloc[:, col_indices]
-        columns = ["time", "fep-lambda", "U", "U_cross"]
-        if ensemble == "npt":
-            columns.append("volume")
         data.columns = columns
         lambda1_col = "fep-lambda"
         data.loc[:, [lambda1_col]] = data[[lambda1_col]].apply(lambda x: round(x, prec))
@@ -634,88 +635,93 @@ def extract_u_nk(
         files, indices=indices, prec=prec, force=force
     )
 
-    if column_lambda2 is None:
+    if column_lambda2 is None:  # No second lambda state value
         u_nk = pd.DataFrame(columns=["time", "fep-lambda"] + lambda_values)
         lc = len(lambda_values)
+        # columns to pull from lammps dump file
         col_indices = [0] + list(columns_lambda1) + [column_U, column_dU]
-    else:
+        # column names from lammps dump file
[…]
